requirement.txt
braintree==3.27.0
defusedxml==0.4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
django-allauth==0.25.2
django-crispy-forms==1.6.0
django-filter==0.14.0
django-localflavor==1.3
django-toolbelt==0.0.1
gunicorn==19.4.5
oauthlib==1.1.1
Pillow==3.2.0
psycopg2==2.6.1
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.1
static3==0.6.1
